[Ref] Move file download to AsyncResourceActions.ts
